requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
Pillow>=8.0.0
tqdm>=4.62.0

